refactor(wallet-sdk): extract the accounts leaf domain (types + predicates) (Phase 2.1)
First slice of the accounts-domain extraction: move the framework-free leaf files into
@agicash/wallet-sdk/src/accounts so the SDK owns the account types/predicates the rest of the
domain (and its consumers) build on.

- account.ts (Account/CashuAccount/SparkAccount types + schemas + predicates
  canSend/ReceiveFromLightning + getAccountBalance), cashu-account.ts (CashuProofSchema, toProof),
  account-cryptography.ts (BIP-85 derivation path) -> packages/wallet-sdk/src/accounts
  (byte-identical git mv; ~/lib/cashu -> @agicash/cashu, ~/lib/money -> @agicash/utils,
  agicash-db/json-models -> @agicash/db-types)
- getAccountHomePath stays in web (it returns app route strings — web-only, meaningless to a
  non-React SDK consumer); the web account.ts shim re-exports the SDK + defines it locally
- old account/cashu-account/account-cryptography paths kept as transitional re-export shims so
  the ~31 consumers stay untouched
- SDK deps: @agicash/{cashu,db-types,utils} (workspace), @agicash/breez-sdk-spark (hoisted to the
  catalog), @cashu/cashu-ts, type-fest, zod

Phase 2. typecheck, biome, build all green; 135 tests; @agicash/wallet-sdk still framework-free.
