+++ apps/web-wallet/app/features/accounts/cashu-account.ts
@@ -1,34 +1,2 @@
-import type { Proof } from '@cashu/cashu-ts';
-import { z } from 'zod/mini';
-import { ProofSchema } from '~/lib/cashu/types';
-
-export const CashuProofSchema = z.object({
-  id: z.string(),
-  accountId: z.string(),
-  userId: z.string(),
-  keysetId: z.string(),
-  amount: z.number(),
-  secret: z.string(),
-  unblindedSignature: z.string(),
-  publicKeyY: z.string(),
-  dleq: ProofSchema.shape.dleq,
-  witness: ProofSchema.shape.witness,
-  state: z.enum(['UNSPENT', 'RESERVED', 'SPENT']),
-  version: z.number(),
-  createdAt: z.string(),
-  reservedAt: z.optional(z.nullable(z.string())),
-  spentAt: z.optional(z.nullable(z.string())),
-});
-
-export type CashuProof = z.infer<typeof CashuProofSchema>;
-
-export const toProof = (proof: CashuProof): Proof => {
-  return {
-    id: proof.keysetId,
-    amount: proof.amount,
-    secret: proof.secret,
-    C: proof.unblindedSignature,
-    dleq: proof.dleq,
-    witness: proof.witness,
-  };
-};
+// Transitional re-export — moved to @agicash/wallet-sdk; removed in the import-cleanup PR.
+export * from '@agicash/wallet-sdk/accounts/cashu-account';
