+++ apps/web-wallet/package.json
@@ -18,7 +18,7 @@
   },
   "dependencies": {
     "@agicash/bc-ur": "0.1.0",
-    "@agicash/breez-sdk-spark": "0.13.5-1",
+    "@agicash/breez-sdk-spark": "catalog:",
     "@agicash/cashu": "workspace:*",
     "@agicash/db-types": "workspace:*",
     "@agicash/opensecret": "catalog:",
