+++ package.json
@@ -5,6 +5,7 @@
   "workspaces": {
     "packages": ["apps/*", "packages/*"],
     "catalog": {
+      "@agicash/breez-sdk-spark": "0.13.5-1",
       "@agicash/opensecret": "0.1.0",
       "@cashu/cashu-ts": "3.6.1",
       "@noble/curves": "1.9.7",
