+++ packages/wallet-sdk/src/accounts/cashu-account.ts
@@ -1,0 +1,34 @@
+import { ProofSchema } from '@agicash/cashu/types';
+import type { Proof } from '@cashu/cashu-ts';
+import { z } from 'zod/mini';
+
+export const CashuProofSchema = z.object({
+  id: z.string(),
+  accountId: z.string(),
+  userId: z.string(),
+  keysetId: z.string(),
+  amount: z.number(),
+  secret: z.string(),
+  unblindedSignature: z.string(),
+  publicKeyY: z.string(),
+  dleq: ProofSchema.shape.dleq,
+  witness: ProofSchema.shape.witness,
+  state: z.enum(['UNSPENT', 'RESERVED', 'SPENT']),
+  version: z.number(),
+  createdAt: z.string(),
+  reservedAt: z.optional(z.nullable(z.string())),
+  spentAt: z.optional(z.nullable(z.string())),
+});
+
+export type CashuProof = z.infer<typeof CashuProofSchema>;
+
+export const toProof = (proof: CashuProof): Proof => {
+  return {
+    id: proof.keysetId,
+    amount: proof.amount,
+    secret: proof.secret,
+    C: proof.unblindedSignature,
+    dleq: proof.dleq,
+    witness: proof.witness,
+  };
+};
